# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Transporte API",
        "description": "Documentación generada con Swagger para la API de transporte",
        "contact": {
            "email": "soporte@transporte.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor Local"
        },
        {
            "url": "https://tu-dominio.com",
            "description": "Servidor Producción"
        }
    ],
    "paths": {
...
        "/api/camioneros": {
            "get": {
                "tags": [
                    "Camioneros"
                ],
                "summary": "Listar camioneros",
                "description": "Obtiene un listado paginado de todos los camioneros junto con sus camiones y paquetes.",
                "operationId": "015139d022ae22eb9a00817b8c379859",
                "responses": {
                    "200": {
                        "description": "Lista de camioneros",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Camionero"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Camioneros"
                ],
                "summary": "Crear un nuevo camionero",
                "operationId": "bd604b782d923e543b032865045869be",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "documento",
                                    "nombre",
                                    "apellido",
                                    "fecha_nacimiento",
                                    "licencia",
                                    "telefono"
                                ],
                                "properties": {
                                    "documento": {
                                        "type": "string",
                                        "example": "123456789"
                                    },
                                    "nombre": {
                                        "type": "string",
                                        "example": "Juan"
                                    },
                                    "apellido": {
                                        "type": "string",
                                        "example": "Pérez"
                                    },
                                    "fecha_nacimiento": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "1985-06-15"
                                    },
                                    "licencia": {
                                        "type": "string",
                                        "example": "C1"
                                    },
                                    "telefono": {
                                        "type": "string",
                                        "example": "3201234567"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Camionero creado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Camionero"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/camioneros/{id}": {
            "get": {
                "tags": [
                    "Camioneros"
                ],
                "summary": "Mostrar un camionero",
                "description": "Devuelve los datos de un camionero específico, incluyendo los camiones asociados.",
                "operationId": "2bfc8c1b9f0ca8d24b20d983bf38bc7c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del camionero",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalles del camionero",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Camionero"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Camioneros"
                ],
                "summary": "Actualizar un camionero",
                "operationId": "290516260c19bb66be553e7619556d11",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del camionero",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "documento": {
                                        "type": "string",
                                        "example": "987654321"
                                    },
                                    "nombre": {
                                        "type": "string",
                                        "example": "Carlos"
                                    },
                                    "apellido": {
                                        "type": "string",
                                        "example": "García"
                                    },
                                    "fecha_nacimiento": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "1990-04-20"
                                    },
                                    "licencia": {
                                        "type": "string",
                                        "example": "C2"
                                    },
                                    "telefono": {
                                        "type": "string",
                                        "example": "3109876543"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Camionero actualizado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Camionero"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Camioneros"
                ],
                "summary": "Eliminar un camionero",
                "operationId": "e463b266bcbcf69b1656e6456ea24efb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del camionero",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Camionero eliminado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Camionero eliminado correctamente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Camionero": {
                "title": "Camionero",
                "description": "Información de un camionero",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "documento": {
                        "description": "Documento de identificación del camionero",
                        "type": "string",
                        "example": "1029384756"
                    },
                    "nombre": {
                        "description": "Nombre del camionero",
                        "type": "string",
                        "example": "Carlos"
                    },
                    "apellido": {
                        "description": "Apellido del camionero",
                        "type": "string",
                        "example": "Pérez"
                    },
                    "fecha_nacimiento": {
                        "description": "Fecha de nacimiento del camionero",
                        "type": "string",
                        "format": "date",
                        "example": "1985-06-20"
                    },
                    "licencia": {
                        "description": "Número de licencia de conducción",
                        "type": "string",
                        "example": "LIC123456"
                    },
                    "telefono": {
                        "description": "Número de teléfono del camionero",
                        "type": "string",
                        "example": "+57 3001234567"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-01-01T12:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-01-01T12:00:00Z"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}